Synchronize entire embeddings calculation phase (#1967) (#1993)
Due to extensive usage of class member fields effectively holding the
state between calls or even phases of processing in the single call
current synchronization in embeddings calculation logic is insufficient
leading to inaccurate results when `add_request` is called in parallel
from multiple threads.

This PR also: 
- adds missing copying, so embeddings calculation phase returns
independent tensor for all supported architectures, not tensors bound to
infer requests that could be overwritten,
- adds additional locks in block manager

This change has already been merged to `releases/2025/1` branch and is
ported as-is.

---------

Co-authored-by: Dariusz Trawinski <Dariusz.Trawinski@intel.com>
Co-authored-by: Ilya Lavrenov <ilya.lavrenov@intel.com>